Be more relaxed with accepting certs in serf
The Desktop product uses neon, so this patch is for Online.
When the WOPI-like host uses a self-signed SSL certificate,
serf refuses to fetch resources from it, so Insert - Image
and Create from Template operations fail.

Related commit: a08552a8c36754930a3268e27aefee70d5cf21df
But in case of serf somehow we do not trigger the interaction
handler, we get an information message box with the text
"Image file cannot be opened"

So I decided to patch serf. The result is the same.
* In Collabora Online 4.0.x we do not check the cert of
  WOPI-like host anyway.
* In Collabora Online 4.2.x we optionally check the cert
  of WOPI-like host. The Create from Template and Insert -
  Image cases will be exceptional, where we won't check it.
  Generally it should not make a big difference.

Change-Id: I3937bf450ef70f3a793ec1c793336d18209b6e62
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96652
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Andras Timar <andras.timar@collabora.com>