lok: Avoid redundant call to UpdateConfigurations
We call doc_setView each time we send an event to LOK such as keystroke
and SidebarController::notifyContextChange event is triggered after that
UpdateConfigurations() call causes layout reallocating for whole sidebar
and we end up keep doing it. It takes significant amount of cpu time for
lok case for no reason.

Signed-off-by: Mert Tumer <mert.tumer@collabora.com>
Change-Id: I8f456156af841fdeaba1934498863a07a7d4554e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120112
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>