lokCalcRTL: h-flip the "combined" tile
This h-flip is needed because the tile containing cell A1 will have the
coordinates (0, 0). Further aim is to have document coordinates for all
tile messages and tile invalidation messages. The LOK client should be
able to map the document coordinates to the correct view coordinates.

In the lok client we also need to h-flip the "regular" tiles at the tile
position.

Change-Id: Iaf89afeb89f916cb490800ee5d34c4780125f3a0