DOCX import: handle a subset of <w:ptab w:alignment="left">
The case when we can map it to a line break. This way the page number is
visually inside the rectangle shape that is behind the field.

The test intentionally uses \n as-is for line-break, because
SwASCWriter::WriteStream() uses \n in the LINEEND_LF case even on
Windows (and not SAL_NEWLINE_STRING), while SAL_NEWLINE_STRING is used
for paragraph-break.

(cherry picked from commit 8bac48991857d222f0e8f0c07b8c4e06649e1632)

Conflicts:
	writerfilter/qa/cppunittests/dmapper/DomainMapper_Impl.cxx

Change-Id: Ic85e57b2391bfac73507727b17240f4d85fc2698
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129205
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>