lokCalcRTL: global RTL: fix chart edit mode rendering
If global RTL flag is set, vcl-window X offset of chart window is
mirrored w.r.t parent window rectangle. This has to be undone to get the
correct chart bounding box in negative X document coordinates so that
the offset calculations for tile rendering remains the same for RTL
documents irrespective of the system/global RTL setting.

Change-Id: I3e1666681af4e7ab1257bcc88d44bbdb053a7d47
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129704
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>