fix Writer key event delayed compression (tdf#145963)
SwEditWin::KeyInput() was stopping the flush timer on every
invocation, which means that as long as key events kept coming
the timer would be started again and again and would not time out,
and so the input buffer would not be actually processed. It seems
that AnyInput(KEYBOARD) there actually does not normally manage
to return true ... except for Win11 where it seems that now
GetQueueStatus() detects even key-up events even though it didn't
before (I've not debugged this enough to be certain).

I think it still makes sense to have the compression in case Writer
get overloaded to help it process the pending input events, so fix
this (but if there are problems, I think an easy solution is simply
to drop it, it doesn't seem to be that necessary). Also decrease
the timeout from 200 to 20ms, as 200ms is a very visible delay.

Change-Id: I760ad9c4f58726a80bd6a9fbf93f2771f10cbdb9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130083
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/130251
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>