lok: ensure unique id for annotations
Problem:
When a cell containing comment is copied and pasted to another cell, the
cloned comment/annotation in the target cell has the same id as the
original one. The lok clients depend upon the id of each comment to
identify them and update coordinates when updates are requested through
.uno:ViewAnnotationsPosition. So the client does not have enough
information to distinguish between comments of same id.

Change-Id: Iebd7281113e0830826aff1bbdaae234bd5d5cd4e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/131689
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>